[Neo] Create neo_utils file for shared functions
